Abstract trig generators into separate class
